Merge

+++ mall-admin/src/main/java/com/macro/mall/service/SmsFlashPromotionProductRelationService.java
@@ -18,7 +18,7 @@ public interface SmsFlashPromotionProductRelationService {
     int create(List<SmsFlashPromotionProductRelation> relationList);
 
     /**
-     * 修改关联相关信息
+     * 修改关联信息
      */
     int update(Long id, SmsFlashPromotionProductRelation relation);
 
@@ -33,7 +33,7 @@ public interface SmsFlashPromotionProductRelationService {
     SmsFlashPromotionProductRelation getItem(Long id);
 
     /**
-     * 分页查询相关商品及促销信息
+     * 分页查询相关商品及限时购促销信息
      *
      * @param flashPromotionId        限时购id
      * @param flashPromotionSessionId 限时购场次id
@@ -42,9 +42,8 @@ public interface SmsFlashPromotionProductRelationService {
 
     /**
      * 根据活动和场次id获取商品关系数量
-     * @param flashPromotionId
-     * @param flashPromotionSessionId
-     * @return
+     * @param flashPromotionId        限时购id
+     * @param flashPromotionSessionId 限时购场次id
      */
-    int getCount(Long flashPromotionId,Long flashPromotionSessionId);
+    long getCount(Long flashPromotionId,Long flashPromotionSessionId);
 }
